requirements.txt
pyTelegramBotAPI==4.27.0
python-dotenv==1.1.1
requests==2.32.4
telebot==0.0.5